static some casts to remove warnings
git-svn-id: https://supercollider.svn.sourceforge.net/svnroot/supercollider/trunk@3991 a380766d-ff14-0410-b294-a243070f3f08

+++ headers/plugin_interface/sc_msg_iter.h
@@ -146,10 +146,10 @@ inline float32 sc_msg_iter::getf(float32 defaultValue)
 			value = OSCfloat(rdpos);
 			rdpos += sizeof(float32);
 		} else if (tags[count] == 'd') {
-			value = static_cast<float64>(OSCdouble(rdpos));
+			value = static_cast<float32>(OSCdouble(rdpos));
 			rdpos += sizeof(float64);
 		} else if (tags[count] == 'i') {
-			value = (float32)OSCint(rdpos);
+			value = static_cast<float32>(OSCint(rdpos));
 			rdpos += sizeof(int32);
 /*		} else if (tags[count] == 's') {
 			value = atof(rdpos);
